third_party/blink/web_tests/http/tests/security/dataURL/xss-DENIED-from-data-url-in-foreign-domain-window-open-expected.txt

CONSOLE MESSAGE: line 1: PASS: Exception thrown successfully.
Opener Frame

Pass: Cross frame access from a data: URL on a different domain was denied.
